requirements.txt
streamlit
chromadb
sentence-transformers
Pillow
google-generativeai
python-dotenv
